<refactor> : <tidying up the project architecture>
