+++ lib/details/model/responses/history_list/history_response.dart
