+++ lib/details/repository/get_history/i_get_history_repo.dart
@@ -1,4 +1,4 @@
-import '../responses/history_list/history_response.dart';
+import '../../model/responses/history_list/history_response.dart';
 
 abstract class GetHistoryRepoInterface {
   Future<HistoryResponse> getHistoryList(String id);
