+++ lib/details/usecases/get_history/get_history_usecase_impl.dart
@@ -1,7 +1,7 @@
-import '../../../shared/model/history_list_view_data.dart';
+import '../../model/history_list_view_data.dart';
+import '../../model/responses/history_list/history_response.dart';
 import '../../repository/get_history/i_get_history_repo.dart';
-import '../../repository/responses/history_list/history_response.dart';
-import '../../shared/mapper/list_history_mapper.dart';
+import '../../utils/mapper/list_history_mapper.dart';
 import 'i_get_history_usecase.dart';
 
 class GetHistoryUsecaseImpl implements GetHistoryUsecaseInterface {
