+++ lib/details/usecases/get_history/i_get_history_usecase.dart
@@ -1,4 +1,4 @@
-import '../../../shared/model/history_list_view_data.dart';
+import '../../model/history_list_view_data.dart';
 
 abstract class GetHistoryUsecaseInterface {
   Future<HistoryListViewData> execute(String id);
