+++ lib/portifolio/controller/providers.dart
@@ -1,30 +1,14 @@
 import 'package:decimal/decimal.dart';
-import 'package:dio/dio.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../shared/model/history_list_view_data.dart';
+import '../../shared/utils/providers/provider.dart';
 import '../endpoint/get_cripto/get_cripto_endpoint_impl.dart';
-import '../endpoint/get_history/get_history_endpoint_impl.dart';
 import '../repository/get_cripto/get_cripto_repo_impl.dart';
-import '../repository/get_history/get_history_repo_impl.dart';
 import '../usecases/get_cripto/get_cripto_usecase_impl.dart';
-import '../usecases/get_history/get_history_usecase_impl.dart';
 import '../usecases/get_total_balance/get_total_balance_usecase_impl.dart';
-
-final dioProvider = Provider(
-  (ref) => Dio(),
-);
-
-// Endpoints
 
 final getCriptoEndpointProvider = StateProvider(
   (ref) => GetCriptoEndpointImpl(
-    ref.watch(dioProvider),
-  ),
-);
-
-final getHistoryEndpointProvider = StateProvider(
-  (ref) => GetHistoryEndpointImpl(
     ref.watch(dioProvider),
   ),
 );
@@ -37,23 +21,11 @@ final getCriptoRepoProvider = StateProvider(
   ),
 );
 
-final getHistoryRepoProvider = StateProvider(
-  (ref) => GetHistoryRepoImpl(
-    ref.watch(getHistoryEndpointProvider),
-  ),
-);
-
 // Usecases
 
 final getCriptoUsecaseProvider = StateProvider(
   (ref) => GetCriptoUsecaseImpl(
     ref.watch(getCriptoRepoProvider),
-  ),
-);
-
-final getHistoryUsecaseProvider = StateProvider(
-  (ref) => GetHistoryUsecaseImpl(
-    ref.watch(getHistoryRepoProvider),
   ),
 );
 
@@ -67,10 +39,6 @@ final getTotalBalanceUsecaseProvider = StateProvider(
 
 final listCriptoProvider = FutureProvider(
   (ref) => ref.watch(getCriptoUsecaseProvider).execute(),
-);
-
-final getHistoryProvider = FutureProvider.family<HistoryListViewData, String>(
-  (ref, id) => ref.watch(getHistoryUsecaseProvider).execute(id),
 );
 
 final getTotalBalanceProvider = FutureProvider.family<Decimal, List<double>>(
