+++ lib/portifolio/endpoint/get_cripto/get_cripto_endpoint_impl.dart
