+++ lib/portifolio/endpoint/get_cripto/i_get_cripto_endpoint.dart
