+++ lib/portifolio/model/cripto_list_view_data.dart
