+++ lib/portifolio/model/cripto_view_data.dart
