+++ lib/portifolio/repository/responses/cripto_coin_list/cripto_response.dart
