+++ lib/portifolio/usecases/get_cripto/i_get_cripto_usecase.dart
@@ -1,4 +1,4 @@
-import '../../../shared/model/cripto_list_view_data.dart';
+import '../../model/cripto_list_view_data.dart';
 
 abstract class GetCriptoUsecaseInterface {
   Future<CriptoListViewData> execute();
