+++ lib/portifolio/usecases/get_total_balance/get_total_balance_usecase_impl.dart
@@ -1,8 +1,8 @@
 import 'package:decimal/decimal.dart';
 
-import '../../../shared/model/cripto_list_view_data.dart';
+import '../../model/cripto_list_view_data.dart';
 import '../../repository/get_cripto/i_get_cripto_repo.dart';
-import '../../shared/mapper/list_cripto_mapper.dart';
+import '../../utils/mapper/list_cripto_mapper.dart';
 import 'i_get_total_balance_usecase.dart';
 
 class GetTotalBalanceUsecaseImpl implements GetTotalBalanceUsecaseInterface {
