+++ lib/portifolio/utils/visible_provider.dart
