+++ lib/portifolio/widgets/cripto_list_tile.dart
@@ -6,9 +6,9 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../details/utils/details_arguments.dart';
-import '../../shared/model/cripto_view_data.dart';
 import '../../shared/utils/utils.dart';
-import '../provider/utils.dart';
+import '../model/cripto_view_data.dart';
+import '../utils/visible_provider.dart';
 
 class CriptoListTile extends HookConsumerWidget {
   const CriptoListTile({
