+++ lib/shared/utils/providers/provider.dart
@@ -1,6 +1,11 @@
+import 'package:dio/dio.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'historico_provider.dart';
+import 'user_amount_provider.dart';
+
+final dioProvider = Provider(
+  (ref) => Dio(),
+);
 
 final userAmountProvider =
     StateNotifierProvider<UserAmountProvider, List<double>>(
