+++ lib/shared/utils/providers/user_amount_provider.dart
